Поменял тип translation в stations.conf

+++ route_editor/include/EditorContext.h
@@ -98,7 +98,7 @@ struct EditorContext
 
     std::map<std::string, ObjectRef> objects_ref;
     RouteMap route_map;
-    std::map<std::string, vsg::vec3> stations_conf;
+    std::map<std::string, vsg::dvec3> stations_conf;
     std::map<std::string, WaypointData> waypoints_conf;
     Topology* topology = nullptr;
